+++ app/components/UI/Stake/components/LearnMoreModal/index.ts
@@ -1,0 +1,1 @@
+export { default as LearnMoreModalFooter } from './LearnMoreModalFooter';
