+++ app/components/UI/Stake/components/PoolStakingLearnMoreModal/index.tsx
@@ -8,19 +8,10 @@ import Text, {
 } from '../../../../../component-library/components/Texts/Text';
 import { ScrollView, View } from 'react-native';
 import BottomSheetHeader from '../../../../../component-library/components/BottomSheets/BottomSheetHeader';
-import BottomSheetFooter, {
-  ButtonsAlignment,
-} from '../../../../../component-library/components/BottomSheets/BottomSheetFooter';
-import {
-  ButtonProps,
-  ButtonSize,
-  ButtonVariants,
-} from '../../../../../component-library/components/Buttons/Button/Button.types';
-import { useNavigation, useRoute, RouteProp } from '@react-navigation/native';
+import { useRoute, RouteProp } from '@react-navigation/native';
 import { POOLED_STAKING_FAQ_URL } from '../../constants';
 import styleSheet from './PoolStakingLearnMoreModal.styles';
 import { useStyles } from '../../../../hooks/useStyles';
-import { MetaMetricsEvents, useMetrics } from '../../../../hooks/useMetrics';
 import useVaultApys from '../../hooks/useVaultApys';
 import InteractiveTimespanChart from './InteractiveTimespanChart';
 import BigNumber from 'bignumber.js';
@@ -30,7 +21,6 @@ import {
 } from './InteractiveTimespanChart/InteractiveTimespanChart.utils';
 import { strings } from '../../../../../../locales/i18n';
 import { parseVaultApyAveragesResponse } from './PoolStakingLearnMoreModal.utils';
-import { EVENT_LOCATIONS, EVENT_PROVIDERS } from '../../constants/events';
 import useVaultApyAverages from '../../hooks/useVaultApyAverages';
 import {
   CommonPercentageInputUnits,
@@ -41,6 +31,7 @@ import { Hex } from 'viem/_types/types/misc';
 import { getDecimalChainId } from '../../../../../util/networks';
 import { endTrace, trace, TraceName } from '../../../../../util/trace';
 import { EARN_EXPERIENCES } from '../../../Earn/constants/experiences';
+import { LearnMoreModalFooter } from '../LearnMoreModal';
 
 interface PoolStakingLearnMoreModalRouteParams {
   chainId: Hex;
@@ -87,10 +78,6 @@ const BodyText = () => {
 
 const PoolStakingLearnMoreModal = () => {
   const { styles } = useStyles(styleSheet, {});
-
-  const { trackEvent, createEventBuilder } = useMetrics();
-
-  const { navigate } = useNavigation();
 
   const route = useRoute<PoolStakingLearnMoreModalRouteProp>();
   const { chainId: routeChainId } = route.params;
@@ -145,42 +132,6 @@ const PoolStakingLearnMoreModal = () => {
     }
   }, [activeTimespanApyAverage, reversedVaultApys]);
 
-  const redirectToLearnMore = () => {
-    navigate('Webview', {
-      screen: 'SimpleWebview',
-      params: {
-        url: POOLED_STAKING_FAQ_URL,
-      },
-    });
-
-    trackEvent(
-      createEventBuilder(MetaMetricsEvents.STAKE_LEARN_MORE_CLICKED)
-        .addProperties({
-          selected_provider: EVENT_PROVIDERS.CONSENSYS,
-          text: 'Learn More',
-          location: EVENT_LOCATIONS.LEARN_MORE_MODAL,
-        })
-        .build(),
-    );
-  };
-
-  const footerButtons: ButtonProps[] = [
-    {
-      variant: ButtonVariants.Secondary,
-      label: strings('stake.learn_more'),
-      size: ButtonSize.Lg,
-      labelTextVariant: TextVariant.BodyMDMedium,
-      onPress: redirectToLearnMore,
-    },
-    {
-      variant: ButtonVariants.Primary,
-      label: strings('stake.got_it'),
-      labelTextVariant: TextVariant.BodyMDMedium,
-      size: ButtonSize.Lg,
-      onPress: handleClose,
-    },
-  ];
-
   const handleTimespanPressed = (numDataPointsToDisplay: number) => {
     setActiveTimespanApyAverage(
       parsedVaultTimespanApyAverages?.[numDataPointsToDisplay],
@@ -227,9 +178,9 @@ const PoolStakingLearnMoreModal = () => {
         )}
         <BodyText />
       </ScrollView>
-      <BottomSheetFooter
-        buttonsAlignment={ButtonsAlignment.Horizontal}
-        buttonPropsArray={footerButtons}
+      <LearnMoreModalFooter
+        onClose={handleClose}
+        learnMoreUrl={POOLED_STAKING_FAQ_URL}
         style={styles.footer}
       />
     </BottomSheet>
